Deadnet_Attacker/requirements.txt
# Core dependencies (cross-platform)
scapy>=2.5.0
netifaces2>=0.0.22
flask>=2.3.0
flask-cors>=4.0.0

# Optional: WebView support (recommended for Windows)
# Install separately if needed: pip install pywebview
pywebview>=4.4.0

# Build tool (only needed for building exe)
pyinstaller>=6.0.0
